requirements.txt
langchain-google-genai
scrapy
requests
Unidecode
google-generativeai
google-genai
google-cloud-aiplatform
langchain
langchain-community
langchain-core
vertexai
faiss-cpu
numpy
Pillow
streamlit
python-dotenv
deepeval